rag-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
# qdrant-client==1.7.0
openai==1.3.5
aiosqlite==0.19.0
asyncpg==0.29.0
# neon-python==0.1.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
# cryptography==41.0.8
cryptography==41.0.7
qdrant-client==1.16.1
